Moved sly-setup fancy to post-init-sly
In master branch when Spacemacs would start, I would get a warning about
undefined sly-init function.

At the same time, the keybindings in lisp mode would be missing.

I moved (sly-setup '(sly-fancy)) to post-init-sly.

Initial test suggests that this removes the warning and the keybindigs are
installed

I also have a diagnostic message in init-sly that will need to be removed.